new project

+++ src/main/java/it/fabrick/meteo/DataJson/GeograJson.java
@@ -3,12 +3,14 @@ package it.fabrick.meteo.DataJson;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class GeograJson {
     private Long istat;
 
-    private Double lng;
+    private BigDecimal lng;
 
-    private Double lat;
+    private BigDecimal lat;
 }
