+++ src/main/java/it/fabrick/meteo/model/GeographicalModel.java
@@ -3,6 +3,8 @@ package it.fabrick.meteo.model;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 
 @Getter
 @Setter
@@ -10,9 +12,9 @@ public class GeographicalModel {
 
     private Long id;
 
-    private Double lng;
+    private BigDecimal lng;
 
-    private Double lat;
+    private BigDecimal lat;
 
 
 }
